backend/requirements.txt
boto3==1.34.0
bcrypt==4.1.2
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0